Cambios en filtros
- Se cambiaron todos los identificadores de Integer a Long.
- Se agregaron métodos para el soporte de búsqueda de actualizaciónes.
- Se arreglaron errores en la implementación previa.
- Se implementó la interfaz JSONSerializable en la jerarquía completa.
- Se modificó el acceso a la base de datos apropiadamente.

+++ src/main/java/com/mypersonalupdates/db/actions/AttributeFilterActions.java
@@ -7,25 +7,28 @@ import org.skife.jdbi.v2.sqlobject.SqlQuery;
 import org.skife.jdbi.v2.sqlobject.SqlUpdate;
 import org.skife.jdbi.v2.sqlobject.customizers.Mapper;
 
+/**
+ * Acciones en la base de datos para la clase {@link com.mypersonalupdates.filters.AttributeFilter}
+ */
 public interface AttributeFilterActions {
     @SqlUpdate("INSERT INTO attribute_filter (ID, providerID, attrID, fieldValue, TYPE) VALUES (:ID, :providerID, :attrID, :fieldValue, :TYPE)")
     int create(
-            @Bind("ID") Integer ID,
-            @Bind("providerID") Integer providerID,
-            @Bind("attrID") Integer attrID,
+            @Bind("ID") Long ID,
+            @Bind("providerID") Long providerID,
+            @Bind("attrID") Long attrID,
             @Bind("fieldValue") String fieldValue,
             @Bind("TYPE") String type
     );
 
     @SqlQuery("SELECT type FROM attribute_filter WHERE ID = :ID")
     String getTypeFromID(
-            @Bind("ID") Integer ID
+            @Bind("ID") Long ID
     );
 
     @SqlQuery("SELECT ID FROM attribute_filter WHERE providerID = :providerID AND attrID = :attrID AND fieldValue = :value AND TYPE = :TYPE LIMIT 1")
-    Integer getIDFromContent(
-            @Bind("providerID") Integer providerID,
-            @Bind("attrID") Integer attrID,
+    Long getIDFromContent(
+            @Bind("providerID") Long providerID,
+            @Bind("attrID") Long attrID,
             @Bind("value") String value,
             @Bind("TYPE") String type
     );
@@ -33,11 +36,11 @@ public interface AttributeFilterActions {
     @SqlQuery("SELECT providerID, attrID FROM attribute_filter WHERE ID = :ID")
     @Mapper(UpdatesProviderAttributeMapper.class)
     UpdatesProviderAttribute getAttrFromID(
-      @Bind("ID") Integer ID
+      @Bind("ID") Long ID
     );
 
     @SqlQuery("SELECT fieldValue FROM attribute_filter WHERE ID = :ID")
     String getFieldValueFromID(
-      @Bind("ID") Integer ID
+      @Bind("ID") Long ID
     );
 }
