+++ src/main/java/com/mypersonalupdates/db/actions/FilterActions.java
@@ -1,27 +1,30 @@
 package com.mypersonalupdates.db.actions;
 
+import com.mypersonalupdates.filters.Filter;
 import org.skife.jdbi.v2.sqlobject.Bind;
 import org.skife.jdbi.v2.sqlobject.GetGeneratedKeys;
 import org.skife.jdbi.v2.sqlobject.SqlQuery;
 import org.skife.jdbi.v2.sqlobject.SqlUpdate;
 
-
+/**
+ * Acciones en la base de datos para la clase {@link Filter}
+ */
 public interface FilterActions {
 
     @SqlUpdate("INSERT INTO filter (TYPE) VALUES (:TYPE)")
     @GetGeneratedKeys
-    Integer create(
+    Long create(
             @Bind("TYPE") String type
     );
 
     @SqlUpdate("DELETE FROM filter WHERE ID = :ID LIMIT 1")
-    Integer remove(
-            @Bind("ID") Integer ID
+    int remove(
+            @Bind("ID") Long ID
     );
 
     @SqlQuery("SELECT TYPE FROM filter WHERE ID = :ID")
     String getTypeFromID(
-            @Bind("ID") Integer ID
+            @Bind("ID") Long ID
     );
 
 }
